+++ src/conn/net_connection.h
@@ -40,14 +40,14 @@ class NetConnection : public Connection {
       : Connection(my_info, peer_info, bid) {}
 
   void SendSetup(SendDevMem** send_dev_mem, void** send_resources,
-                 int buffer_size, ConnInfo* conn_info,
+                 size_t buffer_size, ConnInfo* conn_info,
                  ExchangeConnInfo* ex_info) override;
 
   void RecvSetup(RecvDevMem** recv_dev_mem, void** recv_resources,
-                 int buffer_size, ConnInfo* conn_info,
+                 size_t buffer_size, ConnInfo* conn_info,
                  ExchangeConnInfo* ex_info) override;
 
-  void SendConn(ConnInfo* conn_info, void* send_resources, int buffer_size,
+  void SendConn(ConnInfo* conn_info, void* send_resources, size_t buffer_size,
                 ExchangeConnInfo* peer_ex_info) override;
 
   void RecvConn(ConnInfo* conn_info, void* recv_resources,
